Add PKCE support (#2)
* Add new PKCE setting

Add new setting to enable/disable PKCE feature. A new constant
OIDC_ENABLE_PKCE is available to force the setting's value.

* Store code verifier when creating a new state

Update new state creation method to take an additional parameter
with the PKCE code verifier 's value and store it in the state
value.

* Add method to generate PKCE code challenge

The method will try to generate a code verifier (a random ASCII string)
and a code challenge (SHA256 hash of the verifier) and return an
array with them and the method use to create the code challenge.

If the code verifier generation fails the method will return false.

* Add PKCE code challenge to the authentication url

This is the first step when integrating PKCE into the authentication
workflow. When building the authentication URL a new code verifier
and challenge are created, the code verifier is store in the state
to be accessible at a later stage and the challenge is added as a
query param to the URL along side the method use to generate the
challenge from the verifier.

* Add the verifier to the auth token request

This is the second step when integrating PKCE into the authentication
workflow. Add the code verifier to the auth token request's body.
Code verifier is retieved from the state object created when building
the authentication URL.

* Force boolean type when checking PKCE setting

* First pass at testing the PKCE feature

* Fix wrong number of accepted args

* Fix incorrect merge

---------

Co-authored-by: Clement Boirie <cboirie@beapi.fr>